Refactor: config extraction, modular structure, and upstream merge
Extract configuration to config.yaml
- Add config.yaml with server, paths, catalog URLs, timeouts, heartbeat
- Add app/config.py loader with defaults when file is missing
- All paths and tunables now configurable without code changes

Split monolith into app/ package
- app/routes/: pages (index), api (all /api/* endpoints)
- app/services/: catalog, download, metadata, search, state, zip_utils
- app/templates/: index.html (extracted from inline string)
- app.py becomes thin entry point (~8 lines)

Port upstream _COMPLETE.txt support
- Consolidated translation format with [[PAGE_XXX]] markers
- get_transcription_text, search, save_text support master file
- Backward compatible with individual _pXXX.txt files

README updates
- Document config.yaml options and structure
- Document transcription ZIP (contents, zip_sources)
- Document _COMPLETE.txt format

Code cleanup
- Remove unused imports (Path, metadata)
- Replace print() with logging in catalog
- Remove redundant state.load_metadata_cache wrapper
- Fix duplicate search results when loose + ZIP both have master
- Add __pycache__ to .gitignore

Build: Update PyInstaller workflow for templates and yaml
Made-with: Cursor
