+++ .github/workflows/build.yml
@@ -38,7 +38,7 @@ jobs:
     # ---------------------------------------------
     - name: Build Windows EXE
       if: matrix.os == 'windows-latest'
-      run: pyinstaller --onefile --windowed --icon="rosettaicon.ico" --name "Rosetta Magazine Researcher" app.py
+      run: pyinstaller --onefile --windowed --icon="rosettaicon.ico" --name "Rosetta Magazine Researcher" --add-data "app/templates;app/templates" --hidden-import=yaml app.py
 
     - name: Upload Windows File
       if: matrix.os == 'windows-latest'
@@ -53,7 +53,7 @@ jobs:
     - name: Build Mac
       if: matrix.os == 'macos-latest'
       run: |
-        pyinstaller --onefile --windowed --name "Rosetta_Magazine_Researcher" app.py
+        pyinstaller --onefile --windowed --name "Rosetta_Magazine_Researcher" --add-data "app/templates:app/templates" --hidden-import=yaml app.py
         cd dist
         zip -r Rosetta_Magazine_Researcher_Mac.zip Rosetta_Magazine_Researcher.app
 
@@ -69,7 +69,7 @@ jobs:
     # ---------------------------------------------
     - name: Build Linux
       if: matrix.os == 'ubuntu-latest'
-      run: pyinstaller --onefile --windowed --name "Rosetta_Magazine_Researcher" app.py
+      run: pyinstaller --onefile --windowed --name "Rosetta_Magazine_Researcher" --add-data "app/templates:app/templates" --hidden-import=yaml app.py
 
     - name: Upload Linux File
       if: matrix.os == 'ubuntu-latest'
