+++ app/routes/pages.py
@@ -1,0 +1,10 @@
+"""Page routes."""
+
+from flask import Blueprint, render_template
+
+bp = Blueprint("pages", __name__)
+
+
+@bp.route("/")
+def index():
+    return render_template("index.html")
