+++ app/services/__init__.py
@@ -1,0 +1,5 @@
+"""Business logic and shared state for Rosetta Magazine Researcher."""
+
+from app.services import state
+
+__all__ = ["state"]
